refactor(sandbox): extract harness-runner subprocess from daemon dispatch
